refactor : PathRepository 기능 SectionRepository로 위임 후 삭제
