--- src/main/java/subway/application/repository/PathRepository.java
@@ -1,9 +1,0 @@
-package subway.application.repository;
-
-import subway.domain.Sections;
-
-public interface PathRepository {
-
-    Sections findAllSections();
-
-}
